config docker-compose/swagger

+++ smsender/src/main/java/com/smsender/entity/AccountActions.java
@@ -8,5 +8,5 @@ public interface AccountActions {
 
     Account addingInitialValue();
 
-    Long getBalance();
+    Double getBalance();
 }
